+++ ch.ahoegger.docbox.server/src/main/java/ch/ahoegger/docbox/server/category/CategoryService.java
@@ -3,62 +3,71 @@ package ch.ahoegger.docbox.server.category;
 import java.math.BigDecimal;
 import java.util.Date;
 
+import org.ch.ahoegger.docbox.server.or.app.tables.Category;
+import org.ch.ahoegger.docbox.server.or.app.tables.records.CategoryRecord;
 import org.eclipse.scout.rt.platform.BEANS;
-import org.eclipse.scout.rt.platform.holders.NVPair;
 import org.eclipse.scout.rt.platform.util.StringUtility;
 import org.eclipse.scout.rt.server.jdbc.SQL;
+import org.jooq.Condition;
+import org.jooq.SQLDialect;
+import org.jooq.impl.DSL;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import ch.ahoegger.docbox.or.definition.table.ISequenceTable;
 import ch.ahoegger.docbox.server.document.DocumentCategoryService;
-import ch.ahoegger.docbox.shared.ISequenceTable;
 import ch.ahoegger.docbox.shared.backup.IBackupService;
 import ch.ahoegger.docbox.shared.category.CategoryFormData;
 import ch.ahoegger.docbox.shared.category.CategorySearchFormData;
 import ch.ahoegger.docbox.shared.category.CategoryTableData;
+import ch.ahoegger.docbox.shared.category.CategoryTableData.CategoryTableRowData;
 import ch.ahoegger.docbox.shared.category.ICategoryService;
-import ch.ahoegger.docbox.shared.category.ICategoryTable;
-import ch.ahoegger.docbox.shared.util.SqlFramentBuilder;
 
 /**
  * <h3>{@link CategoryService}</h3>
  *
  * @author Andreas Hoegger
  */
-public class CategoryService implements ICategoryService, ICategoryTable {
+public class CategoryService implements ICategoryService {
+  private static final Logger LOG = LoggerFactory.getLogger(CategoryService.class);
 
   @Override
   public CategoryTableData getTableData(CategorySearchFormData formData) {
-    StringBuilder statementBuilder = new StringBuilder();
-    statementBuilder.append("SELECT ").append(SqlFramentBuilder.columnsAliased(TABLE_ALIAS, CATEGORY_NR, NAME, START_DATE, END_DATE))
-        .append(" FROM ").append(TABLE_NAME).append(" AS ").append(TABLE_ALIAS).append(" ")
-        .append(SqlFramentBuilder.WHERE_DEFAULT);
 
-    // search criteria name
+    Category category = Category.CATEGORY.as("cat");
+
+    Condition condition = DSL.trueCondition();
+    // name
     if (StringUtility.hasText(formData.getName().getValue())) {
-      statementBuilder.append(" AND ").append(SqlFramentBuilder.whereStringContains(NAME, formData.getName().getValue()));
+      condition = condition.and(category.NAME.lower().contains(formData.getName().getValue().toLowerCase()));
     }
-    // seach criteria active
+    // active
     if (formData.getActiveBox().getValue() != null) {
       switch (formData.getActiveBox().getValue()) {
         case TRUE:
-          statementBuilder.append(" AND (").append(SqlFramentBuilder.columnsAliased(TABLE_ALIAS, END_DATE)).append(" >= ").append("CURRENT_DATE")
-              .append(" OR ").append(SqlFramentBuilder.columnsAliased(TABLE_ALIAS, END_DATE)).append(" IS NULL)");
+          condition = condition.and(
+              category.END_DATE.ge(new Date())
+                  .or(category.END_DATE.isNull()));
           break;
         case FALSE:
-          statementBuilder.append(" AND ").append(SqlFramentBuilder.columnsAliased(TABLE_ALIAS, END_DATE)).append(" < ").append("CURRENT_DATE");
+          condition = condition.and(
+              category.END_DATE.lessThan(new Date()));
           break;
       }
     }
 
-    statementBuilder.append(" INTO ")
-        .append(":{td.categoryId}, ")
-        .append(":{td.name}, ")
-        .append(":{td.startDate}, ")
-        .append(":{td.endDate} ");
-
     CategoryTableData tableData = new CategoryTableData();
-    SQL.selectInto(statementBuilder.toString(),
-        new NVPair("td", tableData),
-        formData);
+    DSL.using(SQL.getConnection(), SQLDialect.DERBY)
+        .select(category.CATEGORY_NR, category.NAME, category.START_DATE, category.END_DATE)
+        .from(category)
+        .where(condition)
+        .fetch().forEach(rec -> {
+          CategoryTableRowData row = tableData.addRow();
+          row.setCategoryId(rec.get(category.CATEGORY_NR));
+          row.setName(rec.get(category.NAME));
+          row.setStartDate(rec.get(category.START_DATE));
+          row.setEndDate(rec.get(category.END_DATE));
+        });
     return tableData;
   }
 
@@ -71,11 +80,17 @@ public class CategoryService implements ICategoryService, ICategoryTable {
   @Override
   public CategoryFormData create(CategoryFormData formData) {
     formData.setCategoryId(new BigDecimal(SQL.getSequenceNextval(ISequenceTable.TABLE_NAME)));
-    StringBuilder statementBuilder = new StringBuilder();
-    statementBuilder.append("INSERT INTO ").append(TABLE_NAME);
-    statementBuilder.append(" (").append(SqlFramentBuilder.columns(CATEGORY_NR, NAME, DESCRIPTION, START_DATE, END_DATE)).append(")");
-    statementBuilder.append(" VALUES (:categoryId, :name, :description, :startDate, :endDate )");
-    SQL.insert(statementBuilder.toString(), formData);
+    CategoryRecord category = DSL.using(SQL.getConnection(), SQLDialect.DERBY).newRecord(Category.CATEGORY);
+    category.setCategoryNr(formData.getCategoryId());
+    category.setName(formData.getName().getValue());
+    category.setDescription(formData.getDescription().getValue());
+    if (formData.getStartDate().getValue() != null) {
+      category.setStartDate(new java.sql.Date(formData.getStartDate().getValue().getTime()));
+    }
+    if (formData.getEndDate().getValue() != null) {
+      category.setEndDate(new java.sql.Date(formData.getEndDate().getValue().getTime()));
+    }
+    category.store();
 
     // notify backup needed
     BEANS.get(IBackupService.class).notifyModification();
@@ -85,25 +100,36 @@ public class CategoryService implements ICategoryService, ICategoryTable {
 
   @Override
   public CategoryFormData load(CategoryFormData formData) {
-    StringBuilder statementBuilder = new StringBuilder();
-    statementBuilder.append("SELECT ").append(SqlFramentBuilder.columns(SqlFramentBuilder.columns(NAME, DESCRIPTION, START_DATE, END_DATE)));
-    statementBuilder.append(" FROM ").append(TABLE_NAME);
-    statementBuilder.append(" WHERE ").append(CATEGORY_NR).append(" = :categoryId");
-    statementBuilder.append(" INTO :name, :description, :startDate, :endDate");
-    SQL.selectInto(statementBuilder.toString(), formData);
-    return formData;
+    Category category = Category.CATEGORY.as("cat");
+
+    return DSL.using(SQL.getConnection(), SQLDialect.DERBY)
+        .select(category.NAME, category.DESCRIPTION, category.START_DATE, category.END_DATE)
+        .from(category)
+        .where(category.CATEGORY_NR.eq(formData.getCategoryId()))
+        .fetch()
+        .stream()
+        .map(rec -> {
+          CategoryFormData res = (CategoryFormData) formData.deepCopy();
+          res.getName().setValue(rec.get(category.NAME));
+          res.getDescription().setValue(rec.get(category.DESCRIPTION));
+          res.getStartDate().setValue(rec.get(category.START_DATE));
+          res.getEndDate().setValue(rec.get(category.END_DATE));
+          return res;
+        }).findFirst().orElse(null);
+
   }
 
   @Override
   public CategoryFormData store(CategoryFormData formData) {
-    StringBuilder statementBuilder = new StringBuilder();
-    statementBuilder.append("UPDATE ").append(TABLE_NAME).append(" SET ");
-    statementBuilder.append(NAME).append("= :name, ");
-    statementBuilder.append(DESCRIPTION).append("= :description, ");
-    statementBuilder.append(START_DATE).append("= :startDate, ");
-    statementBuilder.append(END_DATE).append("= :endDate ");
-    statementBuilder.append(" WHERE ").append(CATEGORY_NR).append(" = :categoryId");
-    SQL.update(statementBuilder.toString(), formData);
+    Category category = Category.CATEGORY.as("cat");
+
+    DSL.using(SQL.getConnection(), SQLDialect.DERBY)
+        .update(category)
+        .set(category.NAME, formData.getName().getValue())
+        .set(category.DESCRIPTION, formData.getDescription().getValue())
+        .set(category.START_DATE, formData.getStartDate().getValue())
+        .set(category.END_DATE, formData.getEndDate().getValue())
+        .where(category.CATEGORY_NR.eq(formData.getCategoryId())).execute();
 
     // notify backup needed
     BEANS.get(IBackupService.class).notifyModification();
@@ -112,15 +138,24 @@ public class CategoryService implements ICategoryService, ICategoryTable {
   }
 
   @Override
-  public void delete(Long categoryId) {
-    StringBuilder statementBuilder = new StringBuilder();
-    statementBuilder.append("DELETE FROM ").append(TABLE_NAME).append(" WHERE ").append(CATEGORY_NR).append(" = :categoryId");
-    SQL.delete(statementBuilder.toString(), new NVPair("categoryId", categoryId));
+  public boolean delete(BigDecimal categoryId) {
+    Category category = Category.CATEGORY.as("cat");
+
+    CategoryRecord rec = DSL.using(SQL.getConnection(), SQLDialect.DERBY)
+        .fetchOne(category, category.CATEGORY_NR.eq(categoryId));
+    if (rec == null) {
+      LOG.warn("Try to delete not existing record with id '{}'!", categoryId);
+      return false;
+    }
+    if (rec.delete() != 1) {
+      LOG.error("Could not delete record with id '{}'!", categoryId);
+      return false;
+    }
+
+    BEANS.get(DocumentCategoryService.class).deleteByCategoryId(categoryId);
 
     // notify backup needed
     BEANS.get(IBackupService.class).notifyModification();
-
-    // delete document category connection
-    BEANS.get(DocumentCategoryService.class).deleteByCategoryId(categoryId);
+    return true;
   }
 }
